+++ AzeriteUI/modules/unitframe_playerframe.lua
@@ -176,7 +176,7 @@ end
 local PostCreateAuraButton = function(element, button)
 	
 	-- Downscale factor of the border backdrop
-	local sizeMod = 3/4
+	local sizeMod = 2/4
 
 
 	-- Restyle original elements
@@ -187,8 +187,8 @@ local PostCreateAuraButton = function(element, button)
 	local icon = button.Icon
 	icon:SetTexCoord(5/64, 59/64, 5/64, 59/64)
 	icon:ClearAllPoints()
-	icon:SetPoint("TOPLEFT", 9*sizeMod, -9*sizeMod)
-	icon:SetPoint("BOTTOMRIGHT", -9*sizeMod, 9*sizeMod)
+	icon:SetPoint("TOPLEFT", 3, -3)
+	icon:SetPoint("BOTTOMRIGHT", -3, 3)
 
 	-- Aura stacks
 	local count = button.Count
@@ -227,13 +227,93 @@ local PostCreateAuraButton = function(element, button)
 	})
 	border:SetBackdropBorderColor(Colors.ui.stone[1], Colors.ui.stone[2], Colors.ui.stone[3])
 
+	--[[
+	border:SetBackdrop({
+		bgFile = nil,
+		edgeFile = [=[Interface\ChatFrame\ChatFrameBackground]=],
+		edgeSize = 2,
+		tile = false,
+		tileSize = 0,
+		insets = {
+			left = 0,
+			right = 0,
+			top = 0,
+			bottom = 0
+		}
+	})
+	border:SetBackdropBorderColor(0, 0, 0, 1)
+	
+
+	local bSize = 4
+	local glow2 = border:CreateFrame("Frame")
+	glow2:SetPoint("TOPLEFT", button.Overlay, "TOPLEFT", -bSize, bSize)
+	glow2:SetPoint("BOTTOMRIGHT", button.Overlay, "BOTTOMRIGHT", bSize, -bSize)
+	glow2:SetBackdrop({
+		bgFile = nil, 
+		edgeFile = getPath("border-glow"), 
+		edgeSize = bSize*2,
+		tile = false,
+		tileSize = 0,
+		insets = {
+			left = 0,
+			right = 0,
+			top = 0,
+			bottom = 0
+		}
+	})
+	glow2:SetBackdropBorderColor(0, 0, 0, .75)
+
+	local bSize = 16
+	local glow = glow2:CreateFrame("Frame")
+	glow:SetPoint("TOPLEFT", button.Overlay, "TOPLEFT", -(bSize - 3), (bSize - 3))
+	glow:SetPoint("BOTTOMRIGHT", button.Overlay, "BOTTOMRIGHT", (bSize - 3), -(bSize - 3))
+	--glow:SetPoint("TOPLEFT", button.Overlay, "TOPLEFT", -bSize, bSize)
+	--glow:SetPoint("BOTTOMRIGHT", button.Overlay, "BOTTOMRIGHT", bSize, -bSize)
+	glow:SetBackdrop({
+		bgFile = nil, 
+		edgeFile = getPath("border-glow-overlay"), -- border-glow-overlay
+		edgeSize = bSize*2,
+		tile = false,
+		tileSize = 0,
+		insets = {
+			left = 0,
+			right = 0,
+			top = 0,
+			bottom = 0
+		}
+	})
+	glow:SetBackdropBorderColor(0, 0, 0, .75)
+	]]--
+
 	-- This one we reference, for magic school coloring later on
 	button.Border = border
+	button.Border.Glow = glow
 
 end
 
 -- Anything to post update at all?
 local PostUpdateAuraButton = function(element, button)
+	if (not button) or (not button:IsVisible()) then 
+		return 
+	end 
+	do return end
+
+	if button.isBuff then 
+		--button.Border:SetBackdropBorderColor(0, 0, 0, 1)
+		--button.Border.Glow:SetBackdropBorderColor(0, 0, 0, .75)
+
+		button.Border:SetBackdropBorderColor(1, 0, 0, 1)
+		button.Border.Glow:SetBackdropBorderColor(.7, 0, 0, .5)
+	else
+		local color = button.debuffType and Colors.debuff[button.debuffType]
+		if color then 
+			button.Border:SetBackdropBorderColor(color[1], color[2], color[3], 1)
+			button.Border.Glow:SetBackdropBorderColor(color[1]*.5, color[2]*.5, color[3]*.5, .75)
+		else
+			button.Border:SetBackdropBorderColor(0, 0, 0, 1)
+			button.Border.Glow:SetBackdropBorderColor(0, 0, 0, .75)
+		end
+	end 
 end
 
 
@@ -564,13 +644,14 @@ local Style = function(self, unit, id, ...)
 	-- Auras
 	-----------------------------------------------------------
 	-- not appearing?
+	local aSize, aSpace = 40, 6 -- 42, 4
 	local auras = content:CreateFrame("Frame")
 	auras:Place("BOTTOMLEFT", health, "TOPLEFT", 10, 24)
-	auras:SetSize(42*8 + 8*6, 42) -- auras will be aligned in the available space, this size gives us 7x1 auras
+	auras:SetSize(aSize*8 + aSpace*7, aSize) -- auras will be aligned in the available space, this size gives us 8x1 auras
 
-	auras.auraSize = 40 -- too much?
-	auras.spacingH = 4 -- horizontal/column spacing between buttons
-	auras.spacingV = 4 -- vertical/row spacing between aura buttons
+	auras.auraSize = aSize -- too much?
+	auras.spacingH = aSpace -- horizontal/column spacing between buttons
+	auras.spacingV = aSpace -- vertical/row spacing between aura buttons
 	auras.growthX = "RIGHT" -- auras grow to the left
 	auras.growthY = "UP" -- rows grow downwards (we just have a single row, though)
 	auras.maxVisible = 8 -- when set will limit the number of buttons regardless of space available